feat(view): primitives - protocol & euclidean point
Adds a Primitive protocol that acts as an interface/protocol for static duck
typing that will be common for all primitives(lines, points, polyglons, rectanlges, etc).

Additionally adds a Euclidean point that implements the draw method of the protocol
to returna Python string of the point and also provide a translate method to return a new
Euclidean point.
This point subclasses the NamedTuple class to make it immutable.